requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.5
langchain-text-splitters==0.2.1
openai==1.35.7
faiss-cpu==1.8.0
tiktoken==0.7.0